backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy>=2.0.25
psycopg2-binary==2.9.9
openai>=1.12.0
numpy>=1.24.0

